Changes from livrable1 correction
