+++ domain/image_analysis/opencv_callable/ColorDetector.py
@@ -3,7 +3,7 @@ import numpy as np
 import imutils
 
 from domain.image_analysis.ShapeDetector import ShapeDetector
-from domain.image_analysis.Canny import canny, dilate_mask
+from domain.image_analysis.opencv_callable.Canny import canny, dilate_mask
 
 def color_detector(frame):
     frame = frame.copy()
