+++ domain/image_analysis/opencv_callable/DetectQR.py
