+++ main.py
@@ -1,15 +1,15 @@
 import socketio
 import argparse
 
-from domain.image_analysis.DetectTable import *
+from domain.image_analysis.opencv_callable.DetectTable import *
 from domain.image_path_analysis.ImageToGridConverter import *
 from domain.pathfinding.Astar import Astar
 from test.domain.pathfinding.TestPathFindingImage import test_astar_on_image
-from domain.image_analysis.QR import decode
-from domain.image_analysis.DetectContourPieces import *
-from domain.image_analysis.DetectZoneDep import *
+from domain.image_analysis.opencv_callable.DetectQR import decode
+from domain.image_analysis.opencv_callable.DetectContourPieces import *
+from domain.image_analysis.opencv_callable.DetectZoneDep import *
 from infrastructure.communication_pi.comm_pi import *
-from infrastructure.communication.comm import SendImage, SendText
+from infrastructure.communication_ui.comm_ui import SendImage, SendText
 
 parser = argparse.ArgumentParser(description='Process some integers.')
 parser.add_argument('--d', dest='debug', type=bool, default=False, help='debug mode')
@@ -26,8 +26,9 @@ def pathfinding(path, x, y):
         cv2.circle(test_image.image, (point.j, point.i), 1, [0, 0, 0])
 
     frame = test_image.image
-    SendImage(frame, "optpath")
-    SendImage(frame, "actualpath")
+    comm_ui = Communication_ui()
+    comm_ui.SendImage(frame, "optpath")
+    comm_ui.SendImage(frame, "actualpath")
     cv2.imshow("main", frame)
     cv2.waitKey()
 
@@ -38,8 +39,9 @@ def main_sequence():
     frame = cv2.imread(path)
     obj = decode(frame)
 
-    SendImage(frame, "actualimg")
-    SendText(obj.data, "qrcode")
+    comm_ui = Communication_ui()
+    comm_ui.SendImage(frame, "actualimg")
+    comm_ui.SendText(obj.data, "qrcode")
 
     pathfinding("./image_samples/real_image/globalmonde1QR.jpg", 235, 60)
 
@@ -47,7 +49,8 @@ def main_sequence():
     frame = cv2.imread(path)
     shape = detect_contour_pieces(frame)
 
-    SendImage(shape.frame, "actualimg")
+    comm_ui = Communication_ui()
+    comm_ui.SendImage(shape.frame, "actualimg")
 
     cv2.imshow('EDGES', shape.frame)
     cv2.waitKey()
@@ -58,7 +61,8 @@ def main_sequence():
     frame = cv2.imread(path)
     shape = detect_zone_dep(frame)
 
-    SendImage(shape.frame, "actualimg")
+    comm_ui = Communication_ui()
+    comm_ui.SendImage(shape.frame, "actualimg")
 
     cv2.imshow('EDGES', shape.frame)
     cv2.waitKey()
