+++ test/domain/image_analysis/opencv_callable/ColorDetectorTest.py
@@ -1,4 +1,4 @@
-from domain.image_analysis.DetectZoneDepWorld import *
+from domain.image_analysis.opencv_callable.ColorDetector import *
 from image_samples.real_image import *
 import unittest
 import numpy as np
@@ -6,7 +6,7 @@ import cv2
 import os
 import inspect
 
-class DetectZoneDepWorldTest(unittest.TestCase):
+class ColorDetectorTest(unittest.TestCase):
 
     def setUp(self):
         self.dumb = 0
@@ -15,15 +15,17 @@ class DetectZoneDepWorldTest(unittest.TestCase):
         path = os.path.normpath(os.path.join(os.path.dirname(__file__), os.pardir))
         path = os.path.normpath(os.path.join(path, os.pardir))
         path = os.path.normpath(os.path.join(path, os.pardir))
-        path = os.path.join(path, "./image_samples/real_image/globalmonde.jpg")
+        path = os.path.normpath(os.path.join(path, os.pardir))
+        path = os.path.join(path, "./image_samples/real_image/pieces.jpg")
 
         frame = cv2.imread(path)
 
-        shape = detect_zone_dep_world(frame)
+        shape = color_detector(frame)
+
         cv2.imshow('EDGES', shape.frame)
         cv2.waitKey()
 
-        self.assertEqual(shape.shapes, ['rectangle'])
+        self.assertEqual(shape.shapes, ['circle', 'rectangle', 'circle', 'pentagon'])
 
 if __name__ == '__main__':
     unittest.main()
