+++ test/domain/image_analysis/opencv_callable/DetectContourPiecesTest.py
@@ -1,4 +1,4 @@
-from domain.image_analysis.DetectStartZone import *
+from domain.image_analysis.opencv_callable.DetectContourPieces import *
 from image_samples.real_image import *
 import unittest
 import numpy as np
@@ -6,7 +6,7 @@ import cv2
 import os
 import inspect
 
-class DetectStartZoneTest(unittest.TestCase):
+class DetectContourPiecesTest(unittest.TestCase):
 
     def setUp(self):
         self.dumb = 0
@@ -15,15 +15,16 @@ class DetectStartZoneTest(unittest.TestCase):
         path = os.path.normpath(os.path.join(os.path.dirname(__file__), os.pardir))
         path = os.path.normpath(os.path.join(path, os.pardir))
         path = os.path.normpath(os.path.join(path, os.pardir))
-        path = os.path.join(path, "./image_samples/real_image/globalmonde.jpg")
+        path = os.path.normpath(os.path.join(path, os.pardir))
+        path = os.path.join(path, "./image_samples/real_image/pieces.jpg")
 
         frame = cv2.imread(path)
 
-        shape = detect_start_zone(frame)
+        shape = detect_contour_pieces(frame)
         cv2.imshow('EDGES', shape.frame)
         cv2.waitKey()
 
-        self.assertEqual(shape.shapes, ['square'])
+        self.assertEqual(shape.shapes, ['circle', 'rectangle', 'circle', 'pentagon'])
 
 if __name__ == '__main__':
     unittest.main()
