+++ test/domain/image_analysis/opencv_callable/DetectStartZoneTest.py
@@ -1,4 +1,4 @@
-from domain.image_analysis.ColorDetector import *
+from domain.image_analysis.opencv_callable.DetectStartZone import *
 from image_samples.real_image import *
 import unittest
 import numpy as np
@@ -6,7 +6,7 @@ import cv2
 import os
 import inspect
 
-class ColorDetectorTest(unittest.TestCase):
+class DetectStartZoneTest(unittest.TestCase):
 
     def setUp(self):
         self.dumb = 0
@@ -15,16 +15,16 @@ class ColorDetectorTest(unittest.TestCase):
         path = os.path.normpath(os.path.join(os.path.dirname(__file__), os.pardir))
         path = os.path.normpath(os.path.join(path, os.pardir))
         path = os.path.normpath(os.path.join(path, os.pardir))
-        path = os.path.join(path, "./image_samples/real_image/pieces.jpg")
+        path = os.path.normpath(os.path.join(path, os.pardir))
+        path = os.path.join(path, "./image_samples/real_image/globalmonde.jpg")
 
         frame = cv2.imread(path)
 
-        shape = color_detector(frame)
-
+        shape = detect_start_zone(frame)
         cv2.imshow('EDGES', shape.frame)
         cv2.waitKey()
 
-        self.assertEqual(shape.shapes, ['circle', 'rectangle', 'circle', 'pentagon'])
+        self.assertEqual(shape.shapes, ['square'])
 
 if __name__ == '__main__':
     unittest.main()
