+++ test/domain/image_analysis/opencv_callable/DetectZoneDepWorldTest.py
@@ -1,4 +1,4 @@
-from domain.image_analysis.QR import *
+from domain.image_analysis.opencv_callable.DetectZoneDepWorld import *
 from image_samples.real_image import *
 import unittest
 import numpy as np
@@ -6,7 +6,7 @@ import cv2
 import os
 import inspect
 
-class QRTest(unittest.TestCase):
+class DetectZoneDepWorldTest(unittest.TestCase):
 
     def setUp(self):
         self.dumb = 0
@@ -15,13 +15,16 @@ class QRTest(unittest.TestCase):
         path = os.path.normpath(os.path.join(os.path.dirname(__file__), os.pardir))
         path = os.path.normpath(os.path.join(path, os.pardir))
         path = os.path.normpath(os.path.join(path, os.pardir))
-        path = os.path.join(path, "./image_samples/real_image/qr.png")
+        path = os.path.normpath(os.path.join(path, os.pardir))
+        path = os.path.join(path, "./image_samples/real_image/globalmonde.jpg")
 
         frame = cv2.imread(path)
 
-        obj = decode(frame)
+        shape = detect_zone_dep_world(frame)
+        cv2.imshow('EDGES', shape.frame)
+        cv2.waitKey()
 
-        self.assertEqual(str(obj.data), "b'0-rouge-Zone0'")
+        self.assertEqual(shape.shapes, ['rectangle'])
 
 if __name__ == '__main__':
     unittest.main()
